requirements.txt
aniso8601==7.0.0
dj-database-url==0.5.0
Django==2.2.6
django-environ==0.4.5
django-graphql-jwt==0.1.6
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
xlwt==1.3.0
aniso8601==7.0.0
dj-database-url==0.5.0
Django==2.2.6
django-dotenv==1.4.2
django-environ==0.4.5
django-graphql-jwt==0.1.6
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
xlwt==1.3.0
aniso8601==7.0.0
dj-database-url==0.5.0
Django==2.2.6
django-dotenv==1.4.2
django-environ==0.4.5
django-graphql-jwt==0.1.6
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
xlwt==1.3.0
phonenumbers==8.10.23
django-phonenumber-field==3.0.1